Move Get and bounded types to sp-core
